+++ bank-feeds/RELEASES.md
@@ -677,3 +677,13 @@ Based on:
 - [typescript v3.0.0] bank-feeds
 ### Releases
 - [NPM v3.0.0] https://www.npmjs.com/package/@codat/bank-feeds/v/3.0.0 - bank-feeds
+
+## 2023-12-14 14:06:03
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.126.3 (2.214.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0] bank-feeds
+### Releases
+- [NPM v3.1.0] https://www.npmjs.com/package/@codat/bank-feeds/v/3.1.0 - bank-feeds